de-dupe template
